Simplify bp pangolin
