+++ relays/client-pangolin/Cargo.toml
@@ -16,8 +16,6 @@ codec                  = { package = "parity-scale-codec", version = "2.3.1" }
 
 relay-substrate-client = { path = "../client-substrate" }
 relay-utils            = { path = "../utils" }
-bp-messages            = { path = "../../primitives/messages" }
-bp-runtime             = { path = "../../primitives/runtime" }
 messages-relay         = { path = "../messages" }
 
 ## Bridge dependencies
@@ -27,16 +25,13 @@ messages-relay         = { path = "../messages" }
 #darwinia-bridge-ethereum = { git = "https://github.com/darwinia-network/darwinia-common.git", tag = "v2.8.6" }
 #dp-fee                   = { git = "https://github.com/darwinia-network/darwinia-common.git", tag = "v2.8.6" }
 
+bp-messages            = { path = "../../primitives/messages" }
+bp-runtime             = { path = "../../primitives/runtime" }
+bp-pangolin            = { path = "../../primitives/chain-pangolin" }
+
 ## Substrate Dependencies
 frame-system               = { git = "https://github.com/darwinia-network/substrate.git", branch = "darwinia-v0.12.1" }
 frame-support              = { git = "https://github.com/darwinia-network/substrate.git", branch = "darwinia-v0.12.1" }
 pallet-transaction-payment = { git = "https://github.com/darwinia-network/substrate.git", branch = "darwinia-v0.12.1" }
 sp-core                    = { git = "https://github.com/darwinia-network/substrate.git", branch = "darwinia-v0.12.1" }
 sp-runtime                 = { git = "https://github.com/darwinia-network/substrate.git", branch = "darwinia-v0.12.1" }
-
-
-[dev-dependencies]
-color-eyre   = "0.5"
-futures-test = "0.3"
-tokio        = { version = "1", features = ["time"] }
-codec        = { package = "parity-scale-codec", version = "2.3.1" }
