--- relays/client-pangolin/tests/client_substrate.rs
@@ -1,13 +1,0 @@
-use sp_core::Pair;
-
-#[test]
-fn test_account() {
-    let signer = "//Alice";
-    let pair = sp_core::sr25519::Pair::from_string(&signer, None).unwrap();
-    let public = pair.public();
-    let account = drml_common_primitives::AccountId::from(public.0);
-    assert_eq!(
-        account.to_string(),
-        "5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY"
-    );
-}
